+++ src/main/java/com/woowacamp/soolsool/global/auth/code/AuthErrorCode.java
@@ -1,4 +1,4 @@
-package com.woowacamp.soolsool.core.auth.code;
+package com.woowacamp.soolsool.global.auth.code;
 
 import static org.springframework.http.HttpStatus.BAD_REQUEST;
 
@@ -12,7 +12,7 @@ public enum AuthErrorCode implements ErrorCode {
 
     TOKEN_ERROR(BAD_REQUEST.value(), "A101", "토큰이 유효하지 않습니다."),
     ;
-    
+
     private final int status;
     private final String code;
     private final String message;
