+++ src/main/java/com/woowacamp/soolsool/global/auth/code/AuthResultCode.java
@@ -1,4 +1,4 @@
-package com.woowacamp.soolsool.core.auth.code;
+package com.woowacamp.soolsool.global.auth.code;
 
 import com.woowacamp.soolsool.global.common.ResultCode;
 import lombok.Getter;
