+++ src/main/java/com/woowacamp/soolsool/global/auth/dto/NoAuth.java
@@ -1,4 +1,4 @@
-package com.woowacamp.soolsool.core.auth.dto;
+package com.woowacamp.soolsool.global.auth.dto;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
@@ -6,7 +6,7 @@ import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
 @Retention(RetentionPolicy.RUNTIME)
-@Target(ElementType.PARAMETER)
-public @interface LoginUser {
+@Target({ElementType.TYPE, ElementType.METHOD})
+public @interface NoAuth {
 
 }
