+++ src/test/java/com/woowacamp/soolsool/core/auth/AuthorizationExtractorTest.java
@@ -2,7 +2,7 @@ package com.woowacamp.soolsool.core.auth;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import com.woowacamp.soolsool.core.auth.util.AuthorizationExtractor;
+import com.woowacamp.soolsool.global.auth.util.AuthorizationExtractor;
 import org.apache.http.HttpHeaders;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
