+++ src/test/java/com/woowacamp/soolsool/integration/service/AuthServiceTest.java
@@ -4,9 +4,9 @@ import static com.woowacamp.soolsool.core.member.code.MemberErrorCode.MEMBER_NO_
 import static com.woowacamp.soolsool.core.member.code.MemberErrorCode.MEMBER_NO_MATCH_PASSWORD;
 import static org.assertj.core.api.Assertions.assertThatCode;
 
-import com.woowacamp.soolsool.core.auth.dto.LoginRequest;
-import com.woowacamp.soolsool.core.auth.service.AuthService;
-import com.woowacamp.soolsool.core.auth.util.TokenProvider;
+import com.woowacamp.soolsool.global.auth.dto.LoginRequest;
+import com.woowacamp.soolsool.global.auth.service.AuthService;
+import com.woowacamp.soolsool.global.auth.util.TokenProvider;
 import com.woowacamp.soolsool.global.exception.SoolSoolException;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -17,7 +17,7 @@ import org.springframework.context.annotation.Import;
 @DataJpaTest
 @Import({AuthService.class, TokenProvider.class})
 class AuthServiceTest {
-    
+
     @Autowired
     private AuthService authService;
